+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/DividerElement.tsx
@@ -1,7 +1,8 @@
 import { useAppSelector } from 'app/store/storeHooks';
 import { DividerElementEditMode } from 'features/nodes/components/sidePanel/builder/DividerElementEditMode';
 import { DividerElementViewMode } from 'features/nodes/components/sidePanel/builder/DividerElementViewMode';
-import { selectWorkflowMode, useElement } from 'features/nodes/store/workflowSlice';
+import { useElement } from 'features/nodes/components/sidePanel/builder/use-element';
+import { selectWorkflowMode } from 'features/nodes/store/workflowLibrarySlice';
 import { isDividerElement } from 'features/nodes/types/workflow';
 import { memo } from 'react';
