+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/HeadingElement.tsx
@@ -1,7 +1,8 @@
 import { useAppSelector } from 'app/store/storeHooks';
 import { HeadingElementEditMode } from 'features/nodes/components/sidePanel/builder/HeadingElementEditMode';
 import { HeadingElementViewMode } from 'features/nodes/components/sidePanel/builder/HeadingElementViewMode';
-import { selectWorkflowMode, useElement } from 'features/nodes/store/workflowSlice';
+import { useElement } from 'features/nodes/components/sidePanel/builder/use-element';
+import { selectWorkflowMode } from 'features/nodes/store/workflowLibrarySlice';
 import { isHeadingElement } from 'features/nodes/types/workflow';
 import { memo } from 'react';
