+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/NodeFieldElementStringSettings.tsx
@@ -1,6 +1,6 @@
 import { FormControl, FormLabel, Select } from '@invoke-ai/ui-library';
 import { useAppDispatch } from 'app/store/storeHooks';
-import { formElementNodeFieldDataChanged } from 'features/nodes/store/workflowSlice';
+import { formElementNodeFieldDataChanged } from 'features/nodes/store/nodesSlice';
 import { getDefaultStringOption, type NodeFieldStringSettings, zStringComponent } from 'features/nodes/types/workflow';
 import { omit } from 'lodash-es';
 import type { ChangeEvent } from 'react';
