+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/use-reset-all-node-fields.ts
@@ -1,6 +1,6 @@
 import { useAppStore } from 'app/store/nanostores/store';
 import { fieldValueReset } from 'features/nodes/store/nodesSlice';
-import { selectFormInitialValues, selectNodeFieldElements } from 'features/nodes/store/workflowSlice';
+import { selectFormInitialValues, selectNodeFieldElements } from 'features/nodes/store/selectors';
 import { useCallback } from 'react';
 
 export const useResetAllNodeFields = () => {
