+++ invokeai/frontend/web/src/features/nodes/store/selectors.ts
@@ -1,10 +1,13 @@
 import type { Selector } from '@reduxjs/toolkit';
 import { createSelector } from '@reduxjs/toolkit';
 import type { RootState } from 'app/store/store';
+import { getElement } from 'features/nodes/components/sidePanel/builder/form-manipulation';
 import type { NodesState } from 'features/nodes/store/types';
 import type { FieldInputInstance } from 'features/nodes/types/field';
 import type { AnyNode, InvocationNode, InvocationNodeData } from 'features/nodes/types/invocation';
 import { isBatchNode, isGeneratorNode, isInvocationNode } from 'features/nodes/types/invocation';
+import { isContainerElement, isNodeFieldElement } from 'features/nodes/types/workflow';
+import { uniqBy } from 'lodash-es';
 import { assert } from 'tsafe';
 
 export const selectNode = (nodesSlice: NodesState, nodeId: string): AnyNode => {
@@ -70,7 +73,7 @@ export const selectLastSelectedNodeId = createSelector(selectNodesSlice, ({ node
   return null;
 });
 
-const createNodesSelector = <T>(selector: Selector<NodesState, T>) => createSelector(selectNodesSlice, selector);
+export const createNodesSelector = <T>(selector: Selector<NodesState, T>) => createSelector(selectNodesSlice, selector);
 export const selectNodes = createNodesSelector((nodes) => nodes.nodes);
 export const selectEdges = createNodesSelector((nodes) => nodes.edges);
 export const selectMayUndo = createSelector(
@@ -85,3 +88,41 @@ export const selectMayRedo = createSelector(
 export const selectHasBatchOrGeneratorNodes = createSelector(selectNodes, (nodes) =>
   nodes.filter(isInvocationNode).some((node) => isBatchNode(node) || isGeneratorNode(node))
 );
+
+export const selectWorkflowName = createNodesSelector((nodes) => nodes.name);
+
+export const selectWorkflowId = createNodesSelector((workflow) => workflow.id);
+export const selectWorkflowDescription = createNodesSelector((workflow) => workflow.description);
+export const selectWorkflowNotes = createNodesSelector((workflow) => workflow.notes);
+export const selectWorkflowAuthor = createNodesSelector((workflow) => workflow.author);
+export const selectWorkflowContact = createNodesSelector((workflow) => workflow.contact);
+export const selectWorkflowTags = createNodesSelector((workflow) => workflow.tags);
+export const selectWorkflowVersion = createNodesSelector((workflow) => workflow.version);
+export const selectWorkflowForm = createNodesSelector((workflow) => workflow.form);
+
+export const selectFormRootElementId = createNodesSelector((workflow) => {
+  return workflow.form.rootElementId;
+});
+export const selectFormRootElement = createNodesSelector((workflow) => {
+  return getElement(workflow.form, workflow.form.rootElementId, isContainerElement);
+});
+export const selectIsFormEmpty = createNodesSelector((workflow) => {
+  const rootElement = workflow.form.elements[workflow.form.rootElementId];
+  if (!rootElement || !isContainerElement(rootElement)) {
+    return true;
+  }
+  return rootElement.data.children.length === 0;
+});
+export const selectFormInitialValues = createNodesSelector((workflow) => workflow.formFieldInitialValues);
+export const selectNodeFieldElements = createNodesSelector((workflow) =>
+  Object.values(workflow.form.elements).filter(isNodeFieldElement)
+);
+export const selectWorkflowFormNodeFieldFieldIdentifiersDeduped = createSelector(
+  selectNodeFieldElements,
+  (nodeFieldElements) =>
+    uniqBy(nodeFieldElements, (el) => `${el.data.fieldIdentifier.nodeId}-${el.data.fieldIdentifier.fieldName}`).map(
+      (el) => el.data.fieldIdentifier
+    )
+);
+
+export const buildSelectElement = (id: string) => createNodesSelector((workflow) => workflow.form?.elements[id]);
