+++ invokeai/frontend/web/src/features/nodes/util/workflow/buildWorkflow.ts
@@ -3,8 +3,7 @@ import { useAppStore } from 'app/store/storeHooks';
 import { deepClone } from 'common/util/deepClone';
 import { parseify } from 'common/util/serialize';
 import { selectNodesSlice } from 'features/nodes/store/selectors';
-import type { NodesState, WorkflowsState } from 'features/nodes/store/types';
-import { selectWorkflowSlice } from 'features/nodes/store/workflowSlice';
+import type { NodesState } from 'features/nodes/store/types';
 import { isInvocationNode, isNotesNode } from 'features/nodes/types/invocation';
 import type { WorkflowV3 } from 'features/nodes/types/workflow';
 import { zWorkflowV3 } from 'features/nodes/types/workflow';
@@ -14,12 +13,6 @@ import { useCallback } from 'react';
 import { fromZodError } from 'zod-validation-error';
 
 const log = logger('workflows');
-
-type BuildWorkflowArg = {
-  nodes: NodesState['nodes'];
-  edges: NodesState['edges'];
-  workflow: WorkflowsState;
-};
 
 const workflowKeys = [
   'name',
@@ -35,10 +28,9 @@ const workflowKeys = [
   'form',
 ] satisfies (keyof WorkflowV3)[];
 
-type BuildWorkflowFunction = (arg: BuildWorkflowArg) => WorkflowV3;
-
-export const buildWorkflowFast: BuildWorkflowFunction = ({ nodes, edges, workflow }: BuildWorkflowArg): WorkflowV3 => {
-  const clonedWorkflow = pick(workflow, workflowKeys);
+export const buildWorkflowFast = (nodesState: NodesState): WorkflowV3 => {
+  const { nodes, edges, ...rest } = nodesState;
+  const clonedWorkflow = pick(rest, workflowKeys);
 
   const newWorkflow: WorkflowV3 = {
     ...clonedWorkflow,
@@ -69,9 +61,9 @@ export const buildWorkflowFast: BuildWorkflowFunction = ({ nodes, edges, workflo
   return deepClone(newWorkflow);
 };
 
-export const buildWorkflowWithValidation = ({ nodes, edges, workflow }: BuildWorkflowArg): WorkflowV3 | null => {
+export const buildWorkflowWithValidation = (nodesState: NodesState): WorkflowV3 | null => {
   // builds what really, really should be a valid workflow
-  const workflowToValidate = buildWorkflowFast({ nodes, edges, workflow });
+  const workflowToValidate = buildWorkflowFast(nodesState);
 
   // but bc we are storing this in the DB, let's be extra sure
   const result = zWorkflowV3.safeParse(workflowToValidate);
@@ -91,10 +83,8 @@ export const buildWorkflowWithValidation = ({ nodes, edges, workflow }: BuildWor
 export const useBuildWorkflowFast = (): (() => WorkflowV3) => {
   const store = useAppStore();
   const buildWorkflow = useCallback(() => {
-    const state = store.getState();
-    const { nodes, edges } = selectNodesSlice(state);
-    const workflow = selectWorkflowSlice(state);
-    return buildWorkflowFast({ nodes, edges, workflow });
+    const nodesState = selectNodesSlice(store.getState());
+    return buildWorkflowFast(nodesState);
   }, [store]);
 
   return buildWorkflow;
