+++ invokeai/frontend/web/src/features/ui/components/tabs/WorkflowsTabContent.tsx
@@ -2,7 +2,7 @@ import { ReactFlowProvider } from '@xyflow/react';
 import { useAppSelector } from 'app/store/storeHooks';
 import { ImageViewer } from 'features/gallery/components/ImageViewer/ImageViewer';
 import NodeEditor from 'features/nodes/components/NodeEditor';
-import { selectWorkflowMode } from 'features/nodes/store/workflowSlice';
+import { selectWorkflowMode } from 'features/nodes/store/workflowLibrarySlice';
 import { memo } from 'react';
 
 export const WorkflowsMainPanel = memo(() => {
