+++ invokeai/frontend/web/src/features/workflowLibrary/hooks/useGetFormInitialValues.ts
@@ -1,17 +1,13 @@
 import { useAppStore } from 'app/store/nanostores/store';
+import { getFormFieldInitialValues as _getFormFieldInitialValues } from 'features/nodes/store/nodesSlice';
 import { selectNodesSlice } from 'features/nodes/store/selectors';
-import {
-  getFormFieldInitialValues as _getFormFieldInitialValues,
-  selectWorkflowForm,
-} from 'features/nodes/store/workflowSlice';
 import { useCallback } from 'react';
 
 export const useGetFormFieldInitialValues = () => {
   const store = useAppStore();
 
   const getFormFieldInitialValues = useCallback(() => {
-    const form = selectWorkflowForm(store.getState());
-    const { nodes } = selectNodesSlice(store.getState());
+    const { nodes, form } = selectNodesSlice(store.getState());
     return _getFormFieldInitialValues(form, nodes);
   }, [store]);
